# [FLINK-30709][runtime] NetworkInput#emitNext() should push records to DataOutput within a while loop