@pnowojski 2023-01-18
Generally speaking change looks good to me, but let's wait for the FLINK-30623 to be merged first.