@lindong28 2023-01-26
@1996fanrui Can you help rebase this PR? It will be great to have this improvement in Flink 1.17.