@1996fanrui 2023-01-26
> Thanks for the improvement. Production code LGTM. Could you add a unit test for the `AbstractStreamTaskNetworkInput` or `StreamTaskNetworkInput` (in `StreamTaskNetworkInputTest.java`?), that will check if the records are batched and that batches can be interrupted?

Thanks for your review, I have updated the unit tests.